+++ pkg/objects/log.go
@@ -1,10 +1,1 @@
 package objects
-
-type LogData struct {
-	Logs []LogDatum
-}
-
-type LogDatum struct {
-	DateTime string `comment:"ISO8601"`
-	Message  string
-}
